requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
websockets==13.0
sqlalchemy==2.0.31
jinja2==3.1.3
python-multipart==0.0.9
psycopg[binary]>=3.1.0
python-dotenv==1.0.0
redis>=5.0.0